Add Interface `IExplorerBrowser`. Move its enums there.
